+++ v2/arangodb/client_database.go
@@ -63,6 +63,16 @@ type CreateDatabaseOptions struct {
 	Options CreateDatabaseDefaultOptions `json:"options,omitempty"`
 }
 
+// DatabaseReplicationVersion defines replication protocol version to use for this database
+// Available since ArangoDB version 3.11
+// Note: this feature is still considered experimental and should not be used in production
+type DatabaseReplicationVersion string
+
+const (
+	DatabaseReplicationVersionOne DatabaseReplicationVersion = "1"
+	DatabaseReplicationVersionTwo DatabaseReplicationVersion = "2"
+)
+
 // CreateDatabaseDefaultOptions contains options that change defaults for collections
 type CreateDatabaseDefaultOptions struct {
 	// Default replication factor for collections in database
@@ -71,6 +81,9 @@ type CreateDatabaseDefaultOptions struct {
 	WriteConcern int `json:"writeConcern,omitempty"`
 	// Default sharding for collections in database
 	Sharding DatabaseSharding `json:"sharding,omitempty"`
+	// Replication version to use for this database
+	// Available since ArangoDB version 3.11
+	ReplicationVersion DatabaseReplicationVersion `json:"replicationVersion,omitempty"`
 }
 
 // CreateDatabaseUserOptions contains options for creating a single user for a database.